model_api/requirements.txt
Flask==2.1.1
numpy==1.21.6
Pillow==9.0.1
tensorflow==2.9.1
Werkzeug==2.1.1
Click==8.0.4
itsdangerous==2.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
h5py==3.7.0
Keras-Preprocessing==1.1.2
tensorflow-estimator==2.9.0